+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to GCPy will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - TBD
+### Added
+- Added `create_benchmark_sanity_check_table` routine to `gcpy/benchmark/benchmark_funcs.py` to test if variables are all zero
+or NaN
+
 ## [1.6.1] - 2025-03-24
 ### Added
 - Added GCPy environment file `docs/environment_files/gcpy_environment_py313.yml`, which is based on Python 3.13